+++ src/Tesseract/Pix.cs
@@ -161,6 +161,18 @@ namespace Tesseract
             get { return width; }
         }
 
+        public int XRes
+        {
+            get { return Interop.LeptonicaApi.Native.pixGetXRes(this.handle); }
+            set { Interop.LeptonicaApi.Native.pixSetXRes(this.handle, value); }
+        }
+
+        public int YRes
+        {
+            get { return Interop.LeptonicaApi.Native.pixGetYRes(this.handle); }
+            set { Interop.LeptonicaApi.Native.pixSetYRes(this.handle, value); }
+        }
+
         internal HandleRef Handle
         {
             get { return handle; }